moved zap logic
